artist

+++ Users/src/components/JoinAsArtist.vue
@@ -68,7 +68,7 @@ import Swal from "sweetalert2";
 export default {
   data() {
     return {
-      firstName: "", 
+      firstName: "",
       lastName: "",
       email: "",
       password: "",
@@ -112,28 +112,41 @@ export default {
         );
       } else {
         axios
-          .post("http://localhost:3000/artist/auth/signup", {
-            firstName: this.firstName,
-            lastName: this.lastName,
-            email: this.email,
-            password: this.password,
-            category: this.category,
-          })
+          .get(`http://localhost:3000/artist/auth/${this.email}`)
           .then(({ data }) => {
-            localStorage.setItem("token", data.token);
-            Swal.fire({
-              position: "top-end",
-              icon: "success",
-              title: "Your work has been saved",
-              showConfirmButton: false,
-              timer: 1500,
-            });
+            console.log("mydata=====>", data);
+            if (data.mail === this.email) {
+              swal("Oops!", "Already exist", "error");
+            } else {
+              axios
+                .post("http://localhost:3000/artist/auth/signup", {
+                  firstName: this.firstName,
+                  lastName: this.lastName,
+                  email: this.email,
+                  password: this.password,
+                  category: this.category,
+                })
+                .then(({ data }) => {
+                  localStorage.setItem("token", data.token);
+                  this.$router.push("/Artist-profile");
+                  Swal.fire({
+                    position: "top-end",
+                    icon: "success",
+                    title: "Your work has been saved",
+                    showConfirmButton: false,
+                    timer: 1500,
+                  });
 
-            console.log("registred");
+                  console.log("registred");
+                })
+                .catch((err) => {
+                  console.log(err);
+                  swal("oops", "Something went wrong");
+                });
+            }
           })
           .catch((err) => {
             console.log(err);
-            swal("oops", "Something went wrong");
           });
       }
     },
